Add Java source files and java 8 support
I'm adding the spring-roo generated source files (*.java) and
ignoring the aspectJ ones, it seems to make too much noise.

Unfortunately, spring-roo aims to bring full backward compati-
bility to Java 6 applications, which means the support for
Java 8 does not come out of the box, so dependencies like
hibernate needs to be updated (overriden in the pom.xml file)
in order to use java.time.* and similar packages.